Moved search Input to separate module
